Added refreshing emby by library instead of entire media library. Started path to having optional media servers

+++ service/DeleteWatched.py
@@ -42,22 +42,45 @@ class DeleteWatched:
             
             for user in config['users']:
                 if ('plex_name' in user and 'emby_name' in user):
-                    plex_user_name = user['plex_name']
-                    emby_user_name = user['emby_name']
-                    plex_user_id = self.tautulli_api.get_user_id(plex_user_name)
-                    emby_user_id = self.emby_api.get_user_id(emby_user_name)
+                    plex_user_name = ''
+                    plex_user_id = 0
+                    if 'plex_name' in user:
+                        plex_user_name = user['plex_name']
+                        plex_user_id = self.tautulli_api.get_user_id(plex_user_name)
+                    
+                    emby_user_name = ''
+                    emby_user_id = ''
+                    if 'emby_name' in user:
+                        user_id = self.emby_api.get_user_id(user['emby_name'])
+                        if user_id != self.emby_api.get_invalid_item_id():
+                            emby_user_name = user['emby_name']
+                            emby_user_id = user_id
+                        
+                    
                     if plex_user_id != self.tautulli_api.get_invalid_user_id() and emby_user_id != self.emby_api.get_invalid_item_id():
                         self.user_list.append(UserInfo(plex_user_name, plex_user_id, False, emby_user_name, emby_user_id))
                     else:
                         self.logger.error('{}{}{}: No {}Plex{} user found for {} ... Skipping user'.format(self.service_ansi_code, self.__module__, get_log_ansi_code(), get_plex_ansi_code(), get_log_ansi_code(), plex_user_name))
                         
             for library in config['libraries']:
-                plex_library_name = library['plex_library_name']
-                emby_library_name = library['emby_library_name']
-                plex_library_id = self.tautulli_api.get_library_id(plex_library_name)
-                emby_library_id = self.jellystat_api.get_library_id(emby_library_name)
-                self.libraries.append(LibraryInfo(plex_library_name, plex_library_id, library['plex_media_path'], 
-                                                    emby_library_name, emby_library_id, library['emby_media_path'],
+                plex_library_name = ''
+                plex_library_id = ''
+                plex_media_path = ''
+                if 'plex_library_name' in library and 'plex_media_path' in library:
+                    plex_library_name = library['plex_library_name']
+                    plex_library_id = self.tautulli_api.get_library_id(plex_library_name)    
+                    plex_media_path = library['plex_media_path']
+                    
+                emby_library_name = ''
+                emby_library_id = ''
+                emby_media_path = ''
+                if 'emby_library_name' in library and 'emby_media_path' in library:
+                    emby_library_name = library['emby_library_name']
+                    emby_library_id = self.jellystat_api.get_library_id(emby_library_name)
+                    emby_media_path = library['emby_media_path']
+                
+                self.libraries.append(LibraryInfo(plex_library_name, plex_library_id, plex_media_path, 
+                                                    emby_library_name, emby_library_id, emby_media_path,
                                                     library['utilities_path']))
 
             self.delete_time_hours = config['delete_time_hours']
@@ -75,14 +98,15 @@ class DeleteWatched:
         try:
             dateTimeStringForHistory = get_datetime_for_history_plex_string(1)
             for user in self.user_list:
-                watchedItems = self.tautulli_api.get_watch_history_for_user_and_library(user.plex_user_id, lib.plex_library_id, dateTimeStringForHistory)
-                for item in watchedItems:
-                    if item['watched_status'] == 1:
-                        fileName = self.tautulli_api.get_filename(item['rating_key'])
-                        if len(fileName) > 0:
-                            hoursSincePlay = self.hours_since_play(False, datetime.fromtimestamp(item['stopped']))
-                            if hoursSincePlay >= self.delete_time_hours:
-                                returnFileNames.append(DeleteFileInfo(fileName.replace(lib.plex_media_path, lib.utilities_path), user.plex_user_name, 'Plex', get_plex_ansi_code()))
+                if user.plex_user_id != 0 and lib.plex_library_id != '' and lib.plex_media_path != '':
+                    watchedItems = self.tautulli_api.get_watch_history_for_user_and_library(user.plex_user_id, lib.plex_library_id, dateTimeStringForHistory)
+                    for item in watchedItems:
+                        if item['watched_status'] == 1:
+                            fileName = self.tautulli_api.get_filename(item['rating_key'])
+                            if len(fileName) > 0:
+                                hoursSincePlay = self.hours_since_play(False, datetime.fromtimestamp(item['stopped']))
+                                if hoursSincePlay >= self.delete_time_hours:
+                                    returnFileNames.append(DeleteFileInfo(fileName.replace(lib.plex_media_path, lib.utilities_path), user.plex_user_name, 'Plex', get_plex_ansi_code()))
 
         except Exception as e:
             self.logger.error("{}{}{}: Find {}Plex{} watched media {}error={}{}.".format(self.service_ansi_code, self.__module__, get_log_ansi_code(), get_plex_ansi_code(), get_log_ansi_code(), get_tag_ansi_code(), get_log_ansi_code(), e))
@@ -92,23 +116,24 @@ class DeleteWatched:
     def find_emby_watched_media(self, lib):
         returnFileNames = []
         try:
-            watchedItems = self.jellystat_api.get_library_history(lib.emby_library_id)
-            for item in watchedItems:
-                for user in self.user_list:
-                    if item['UserName'] == user.emby_user_name:
-                        item_id = '0'
-                        if 'EpisodeId' in item and item['EpisodeId'] is not None:
-                            item_id = item['EpisodeId']
-                        else:
-                            item_id = item['NowPlayingItemId']
-                            
-                        if self.emby_api.get_watched_status(user.emby_user_name, item_id) == True:
-                            hoursSincePlay = self.hours_since_play(True, datetime.fromisoformat(item['ActivityDateInserted']))
-                            if hoursSincePlay >= self.delete_time_hours:
-                                emby_item = self.emby_api.search_item(item_id)
-                                if emby_item is not None:
-                                    returnFileNames.append(DeleteFileInfo(emby_item['Path'].replace(lib.emby_media_path, lib.utilities_path)), user.emby_user_name, 'Emby', get_emby_ansi_code())
-                        break
+            if (lib.emby_library_id != '' and lib.emby_media_path != ''):
+                watchedItems = self.jellystat_api.get_library_history(lib.emby_library_id)
+                for item in watchedItems:
+                    for user in self.user_list:
+                        if item['UserName'] == user.emby_user_name:
+                            item_id = '0'
+                            if 'EpisodeId' in item and item['EpisodeId'] is not None:
+                                item_id = item['EpisodeId']
+                            else:
+                                item_id = item['NowPlayingItemId']
+
+                            if self.emby_api.get_watched_status(user.emby_user_name, item_id) == True:
+                                hoursSincePlay = self.hours_since_play(True, datetime.fromisoformat(item['ActivityDateInserted']))
+                                if hoursSincePlay >= self.delete_time_hours:
+                                    emby_item = self.emby_api.search_item(item_id)
+                                    if emby_item is not None:
+                                        returnFileNames.append(DeleteFileInfo(emby_item['Path'].replace(lib.emby_media_path, lib.utilities_path)), user.emby_user_name, 'Emby', get_emby_ansi_code())
+                            break
         
         except Exception as e:
             self.logger.error("{}{}{}: Find {}Emby{} watched media {}error={}{}.".format(self.service_ansi_code, self.__module__, get_log_ansi_code(), get_emby_ansi_code(), get_log_ansi_code(), get_tag_ansi_code(), get_log_ansi_code(), e))
@@ -137,15 +162,27 @@ class DeleteWatched:
         # If shows were deleted clean up folders and notify
         if number_of_deleted_media > 0:
             try:
+                plex_lib_refreshed = False
+                emby_lib_refreshed = False
+                
                 # Notify Plex to refresh
                 self.plex_api.switch_plex_account_admin()
                 for lib in self.libraries:
-                    self.plex_api.set_library_scan(lib.plex_library_name)
+                    if lib.plex_library_name != '':
+                        self.plex_api.set_library_scan(lib.plex_library_name)
+                        plex_lib_refreshed = True
+                        
+                    if lib.emby_library_id != '':
+                        self.emby_api.set_library_scan(lib.emby_library_id)
+                        emby_lib_refreshed = True
                 
-                # Notify Emby to refresh
-                self.emby_api.set_library_scan()
-                
-                self.logger.info("{}{}{}: Notifying Media Servers to Refresh".format(self.service_ansi_code, self.__module__, get_log_ansi_code()))
+                if plex_lib_refreshed == True and emby_lib_refreshed == True:
+                    self.logger.info('{}{}{}: Notified {}Plex{} and {}Emby{} to refresh'.format(self.service_ansi_code, self.__module__, get_log_ansi_code(), get_plex_ansi_code(), get_log_ansi_code(), get_emby_ansi_code(), get_log_ansi_code()))
+                elif plex_lib_refreshed == True:
+                    self.logger.info('{}{}{}: Notified {}Plex{} to refresh'.format(self.service_ansi_code, self.__module__, get_log_ansi_code(), get_plex_ansi_code(), get_log_ansi_code()))
+                elif emby_lib_refreshed == True:
+                    self.logger.info('{}{}{}: Notified {}Emby{} to refresh'.format(self.service_ansi_code, self.__module__, get_log_ansi_code(), get_emby_ansi_code(), get_log_ansi_code()))
+
             except Exception as e:
                 self.logger.error("{}{}{}: Clean up failed {}error={}{}.".format(self.service_ansi_code, self.__module__, get_log_ansi_code(), get_tag_ansi_code(), get_log_ansi_code(), e))
